requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
pywhatkit==5.4.7 